+++ types/protocol-mapping.d.ts
@@ -5311,6 +5311,20 @@ export namespace ProtocolMapping {
             paramsType: [Protocol.BluetoothEmulation.SimulateGATTOperationResponseRequest];
             returnType: void;
         };
+        /**
+         * Adds a service with |uuid| to the peripheral with |address|.
+         */
+        'BluetoothEmulation.addService': {
+            paramsType: [Protocol.BluetoothEmulation.AddServiceRequest];
+            returnType: Protocol.BluetoothEmulation.AddServiceResponse;
+        };
+        /**
+         * Removes the service respresented by |id| from the peripheral with |address|.
+         */
+        'BluetoothEmulation.removeService': {
+            paramsType: [Protocol.BluetoothEmulation.RemoveServiceRequest];
+            returnType: void;
+        };
     }
 }